+++ websieve/99_outro.js
@@ -1,1 +1,1 @@
-})(window);
+})(window, /*@__NOINLINE__*/ (() => "fetch-util:standalone-api:v1")());